Score data class, returned by the get matching items method, and stored in search activity as each item

+++ app/src/main/java/com/example/lostandfound/MatchingFunctions/BrandScore.kt
@@ -1,7 +1,7 @@
 package com.example.lostandfound.MatchingFunctions
 
-fun getBrandScore(brand1: String, brand2: String): Int{
+fun getBrandScore(brand1: String, brand2: String): Double{
     // if they are classified as similar using the word comparison algorithm, return 3
     // else 0
-    return if (areWordsSimilar(brand1, brand2)) 3 else 0
+    return if (areWordsSimilar(brand1, brand2)) 3.0 else 0.0
 }
