+++ app/src/main/java/com/example/lostandfound/MatchingFunctions/ColorScore.kt
@@ -4,7 +4,7 @@ import com.example.lostandfound.Data.FoundItem
 import com.example.lostandfound.Data.LostItem
 import kotlin.math.min
 
-fun getColorScore(lostColors: List<String>, foundColors: List<String>): Int{
+fun getColorScore(lostColors: List<String>, foundColors: List<String>): Double{
     // convert the lists to sets
     val lostSet = lostColors.toSet()
     val foundSet = foundColors.toSet()
@@ -13,5 +13,5 @@ fun getColorScore(lostColors: List<String>, foundColors: List<String>): Int{
     // so that subsets would still give a full score of 3
     // as some people may recognise an object as red, while others red and blue.
     val numIntersection = lostSet.intersect(foundSet).size
-    return 3*numIntersection / min(lostSet.size, foundSet.size)
+    return 3*numIntersection / min(lostSet.size, foundSet.size).toDouble()
 }
